requirements.txt
python-dotenv
beautifulsoup4
python3-scrapy